+++ core/store/orm/config.go
@@ -264,6 +264,17 @@ func (c Config) DatabaseMaximumTxDuration() time.Duration {
 	return c.getWithFallback("DatabaseMaximumTxDuration", parseDuration).(time.Duration)
 }
 
+// DatabaseBackupEnabled turns on the database backup on node start if set to true
+func (c Config) DatabaseBackupEnabled() bool {
+	return c.getWithFallback("DatabaseBackupEnabled", parseBool).(bool)
+}
+
+// DatabaseBackupFrequency turns on the periodic database backup if set to a positive value
+// DatabaseBackupEnabled must be then set to true as well
+func (c Config) DatabaseBackupFrequency() time.Duration {
+	return c.getWithFallback("DatabaseBackupFrequency", parseDuration).(time.Duration)
+}
+
 // DatabaseTimeout represents how long to tolerate non response from the DB.
 func (c Config) DatabaseTimeout() models.Duration {
 	return models.MustMakeDuration(c.getWithFallback("DatabaseTimeout", parseDuration).(time.Duration))